+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
-## v2.4.0 - unreleased
+## v2.3.1 - unreleased
 
 ### Added
 
 - `generate_no_rewind()` and `generate_core_no_rewind()` to `Scru128Generator`
-- `rollback_allowance` parameter to `generate_core()` of `Scru128Generator`
+  (experimental)
 
 ## v2.3.0 - 2022-12-25